+++ src/components/ES6/scroll-reveal/js/index.js
@@ -21,7 +21,7 @@ export const SCROLL_REVEAL = ( ( module, $, window, document ) => {
 	
 	
     module.SCROLL_REVEAL               = module.SCROLL_REVEAL || {};
-    module.SCROLL_REVEAL.version       = '0.1.0';
+    module.SCROLL_REVEAL.version       = '0.1.2';
     module.SCROLL_REVEAL.documentReady = function( $ ) {
 
 		
@@ -38,7 +38,7 @@ export const SCROLL_REVEAL = ( ( module, $, window, document ) => {
 						"from"     : {"opacity":0,"x":70},
 						"to"       : {"opacity":1,"x":0},
 						"ease"     : "Power2.easeOut",
-						"duration" : 0.8,
+						"duration" : 0.4,
 						"delay"    : 0,
 						"infinite" : false,
 						"viewport" : '100%' //A percentage of the viewport's height.
@@ -57,6 +57,19 @@ export const SCROLL_REVEAL = ( ( module, $, window, document ) => {
 				//A percentage of the viewport's height.
 				viewport = config.viewport;
 				
+                
+                if ( typeof viewport === typeof undefined ) viewport = '100%';
+                if ( typeof myEase === typeof undefined ) myEase = 'Power2.easeOut';
+                if ( typeof myDelay === typeof undefined ) myDelay = 0;
+                if ( typeof myDuration === typeof undefined ) myDuration = 0.4;
+                if ( typeof infinite === typeof undefined ) infinite = false;
+
+
+                //Return a value
+                if ( type == 'viewport' ) return viewport;
+                if ( type == 'delay' ) return myDelay;
+                if ( type == 'loop' ) return ( infinite ) ? 1 : 0;
+
 				
 				if( Object.prototype.toString.call( fromCSS ) == '[object String]' ) {
 					//Add class when element becomes visible
@@ -68,7 +81,14 @@ export const SCROLL_REVEAL = ( ( module, $, window, document ) => {
 					if ( type == 'from-anim' ) obj.removeClass( toCSS );
 
 					//Target animation
-					if ( type == 'to' ) obj.addClass( toCSS );
+                    if ( type == 'to' ) {
+
+                        setTimeout( function(){
+                            obj.addClass( toCSS );
+                        }, myDelay*1000 );
+
+                    }
+
 
 					
 				} else {
@@ -102,16 +122,7 @@ export const SCROLL_REVEAL = ( ( module, $, window, document ) => {
 
 					
 				}
-				
-				//Reversing Scroll Animations for Loop  
-				if ( type == 'loop' ) {
-					if ( infinite ) {
-						return 1;
-					} else {
-						return 0;
-					}	
-				}
-
+	
 				
 
 			};